Fix AbstractComponent Single-run tasks.
This commit ensures that tasks that have been delayed to run even after the component has been unloaded are made sure to stop. Prior to this the tasks were removed as a means of stopping it from running twice in the case of it being called to run twice in one session, to be revisited in the near future.