Add ComponentState to BlurSession.
This commit adds a `state` field to BlurSession for adding the typical ComponentState values to BlurSession without BlurSession actually being of type AbstractComponent. Furthermore, the `enabled` boolean has been removed and instead `state` is used to dictate whether the session is enabled.

This commit also adds a new `load` method to BlurSession for further control over stages of the session starting up.