Fix First World Problems.
This commit switches the modules ListMultimap in ModuleManager to use the LinkedList implementation instead of ArrayList. This fixes the order of module loading to the order of insertion. For the past 18 months, modules that have been implemented have mostly not depended on each other so this never stood out, even in the ubiquitous debug lines of modules loading, I failed to spot this bug. Finally this bug now fixes many errors caused through module dependencies (I've yet to implement to actual module dependencies feature).

ArrayList implementation uses HashMap for it's primary collection storage, that is, HashMap<K, ArrayList<V>>. The problem is the keys (which are Class) are not in order, so when values() is fetched, it fetches the collated ArrayList values in that Class hashed order.

WOOOOOOOOOOOOOOOOOW BABY!