Add quit boolean param to BlurSession#removePlayer(BlurPlayer)`
This commit changes the behaviour of players removal from sessions. Essentially, if a player is being removed from the session due to the session stopping, they are not considered to be quitting, altering behaviour via the `nextSession` variable in the method, which is passed to `PlayerLeaveSessionEvent` invocation. Furthermore, `nextSession` is used in event listeners to determine following behaviour based on whether it is null.