Improve player quitting logic quality and rational.
BlurSessionListener has been moved into RootBlurSession, effectively calling the equivalent code in one Listener.

On PlayerQuitEvent, the root session will recursively remove the player from their session, and continiously climb in session ancestry to the root to remove the player from it.

It is important to note nextSession now updates to the player's parent session when the player is quitting. This is a valid change and exists to assist with cleaning up players on any subsequent events triggered by PlayerQuitEvent.

Albeit not a perfect method, this will probably clear all cases that involve cleaning the player up. However, in the future it might be better to accurately track which sessions the player is actually a part of and specifically invoke removePlayer on those sessions (currently calling removePlayer when a player isnt in that session does nothing, so no harm done anyways).