Make Blur#getPlayer(Player) nullable, no longer creating instances.
Important change here, getPlayer can be called durign events that are called right after the player leaves. This has happened in production where a EntityDamageByBlockEvent was called right after the player left the game. This creates inconsistencies as Blur destroys the player as they leave, but the event is called assuming the player is online.